refactor(): improvements and update imports
